changed the file-location and file dependency's
