+++ safe_video/__init__.py
@@ -1,2 +1,2 @@
 from .ui import UI_App
-from .number_plate_recognition.plate_recognition import ObjectDetection
+from .number_plate_recognition import *
