+++ safe_video/number_plate_recognition/__init__.py
@@ -1,1 +1,2 @@
 from .plate_recognition import ObjectDetection
+from .utils import *
